pyproject.toml
[project]
name = "robotic-vision-phase1"
version = "0.1.0"
description = "Phase 1 baseline pipeline for household object detection."
requires-python = ">=3.11,<3.13"
dependencies = [
    "jsonschema>=4.25.0",
    "pandas>=2.2.0",
    "pydantic>=2.11.0",
    "pyyaml>=6.0.0",
    "typer>=0.16.0",
]

[project.optional-dependencies]
ml = [
    "datasets>=4.0.0",
    "pillow>=11.0.0",
    "pycocotools>=2.0.10",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.0",
    "ultralytics>=8.3.0",
]
dev = [
    "pytest>=8.4.0",
]

[project.scripts]
phase1 = "src.cli.main:run"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
